requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
funcsigs==1.0.2
more-itertools==5.0.0
numpy==1.16.0
pathlib2==2.3.3
ped-parser==1.6.6
pluggy==0.8.1
py==1.7.0
pysam==0.8.4
pytest==4.1.1
PyVCF==0.6.8
scandir==1.9.0
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0